delete unused files, reorganize things
