+++ frontend/src/metabase/admin/datamodel/components/MetricItem.jsx
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import PropTypes from "prop-types";
 
 import Icon from "metabase/components/Icon";
-import ObjectActionSelect from "../ObjectActionSelect";
+import ObjectActionSelect from "./ObjectActionSelect";
 
 import * as Q_DEPRECATED from "metabase/lib/query";
