+++ frontend/src/metabase/admin/datamodel/containers/SegmentListApp.jsx
@@ -2,13 +2,12 @@ import React from "react";
 import { t } from "ttag";
 
 import Segment from "metabase/entities/segments";
-import SegmentItem from "metabase/admin/datamodel/components/database/SegmentItem";
+import SegmentItem from "metabase/admin/datamodel/components/SegmentItem";
+import FilteredToUrlTable from "metabase/admin/datamodel/hoc/FilteredToUrlTable";
 
 import Button from "metabase/components/Button";
 import Link from "metabase/components/Link";
 import { DatabaseSchemaAndTableDataSelector } from "metabase/query_builder/components/DataSelector";
-
-import FilteredToUrlTable from "../hoc/FilteredToUrlTable";
 
 @Segment.loadList({ wrapped: true })
 @FilteredToUrlTable("segments")
