+++ frontend/src/metabase/admin/datamodel/hoc/withTableMetadataLoaded.js
